I think the code could be simpler
- Triming the FilteredContext-related features
- Moving to a core API on Actor and adding behaviors though data

And what's missing from my comments is that I think the input parsing is
the key to the before_hooks and after_hooks approach. Instead of
overloading `before` and `after` with specific procedures walking inputs
and outputs to behave correctly, each call to input could parse its
options and add the appropriate hooks.

Instead of extending the Actor, we can build some kind of a pipeline of
collaborators that try to read the input definition and build the proper
hooks.

I would happily explore all that pairing if you're interested ;-)